+++ tests/test_all_fields.py
@@ -1,13 +1,8 @@
-from collections import defaultdict
 from random import choice
-import json
-import os
 import unittest
 from tap_tester.base_suite_tests.all_fields_test import AllFieldsTest
-from tap_tester import menagerie, runner, connections
+from tap_tester import menagerie, runner
 from tap_tester.logger import LOGGER
-from datetime import datetime as dt
-from datetime import timedelta
 from base import GA4Base
 
 
@@ -18,81 +13,49 @@ class GA4AllFieldsTest(AllFieldsTest, GA4Base):
     METRIC = 'METRIC'
     fields_1 = None
     fields_2 = None
-
+    field_exclusions = {}
 
     @staticmethod
     def name():
         return "tt_ga4_all_fields"
 
-
     def streams_to_test(self):
         # testing all streams creates massive quota issues
-        custom_id_1 = self.custom_reports_names_to_ids()['Test Report 1']
-        custom_id_2 = self.custom_reports_names_to_ids()['Test Report 2']
-        return {
-            custom_id_1,
-            custom_id_2
-        }
+        return {'Test Report 1', 'Test Report 2'}
 
-
-    def streams_to_selected_fields(self):
+    def streams_to_selected_fields(self):  # pylint: disable=arguments-differ
+        # TODO - BUG https://jira.talendforge.org/browse/TDL-21467
         if not self.fields_1 and not self.fields_2:
-            self.fields_1 = self.select_random_fields()
-            self.fields_2 = self.select_random_fields()
+            self.fields_1 = self.select_random_fields() - {"nth_hour"}
+            self.fields_2 = self.select_random_fields() - {"nth_hour"}
         return {
-            "Test Report 1": self.fields_1,
-            "Test Report 2": self.fields_2
+            "Test Report 1": self.fields_1 | self.expected_automatic_fields()["Test Report 1"],
+            "Test Report 2": self.fields_2 | self.expected_automatic_fields()["Test Report 2"]
         }
 
+    @property
+    def schemas(self):
+        test_catalogs = [catalog for catalog in self.found_catalogs
+                         if catalog.get('stream_name') in self.streams_to_test()]
+        return {
+            catalog['stream_name']: menagerie.get_annotated_schema(
+                self.conn_id, catalog['stream_id'])
+            for catalog in test_catalogs}
 
     ##########################################################################
-    ### Overriden setup
+    # Overridden setup
     ##########################################################################
 
-
-    # Set up requires and extra step to get catalogs for selecting random fields
-    def setUp(self):
-        """
-        Setup for tests in this module.
-        """
-        cls = type(self)
-
-        if cls.synced_records:
-            return
-
-        # instantiate connection
-        conn_id = connections.ensure_connection(self, original_properties=True)
-
-        # run check mode
-        found_catalogs = self.run_and_verify_check_mode(conn_id)
-
-        # table and field selection
-        test_catalogs = [catalog for catalog in found_catalogs
-                         if catalog.get('tap_stream_id') in self.streams_to_test()]
-
-        # For taps that need random selection of fields
-        self.schemas = {catalog['stream_name']: menagerie.get_annotated_schema(conn_id, catalog['stream_id'])
-                        for catalog in test_catalogs}
-
-        self.select_streams_and_fields(conn_id, test_catalogs, self.streams_to_selected_fields())
-        self.selected_fields = self.get_all_selected_fields_from_metadata(conn_id, test_catalogs)
-
-        # run initial sync
-        cls.record_count_by_stream = self.run_and_verify_sync_mode(conn_id)
-        cls.synced_records = runner.get_records_from_target_output()
-
-
     ##########################################################################
-    ### Helper methods
+    # Helper methods
     ##########################################################################
-
 
     def get_field_exclusions(self):
         field_exclusions = {'DIMENSION': {}, 'METRIC': {}}
         for field in self.schemas['Test Report 1']['metadata']:
             behavior = field['metadata'].get('behavior')
 
-            if field['breadcrumb'] == [] or (behavior != 'DIMENSION' and behavior != 'METRIC'):
+            if field['breadcrumb'] == [] or behavior not in ('DIMENSION', 'METRIC'):
                 continue
 
             field_name = field['breadcrumb'][1]
@@ -100,8 +63,8 @@ class GA4AllFieldsTest(AllFieldsTest, GA4Base):
 
         self.field_exclusions = field_exclusions
 
-
-    def exclude_field(self, all_selectable_dimensions, all_selectable_metrics, selected_field, selected_field_type):
+    def exclude_field(self, all_selectable_dimensions, all_selectable_metrics,
+                      selected_field, selected_field_type):
         excluded_fields = self.field_exclusions[selected_field_type][selected_field]
 
         if selected_field_type == self.DIMENSION:
@@ -116,7 +79,6 @@ class GA4AllFieldsTest(AllFieldsTest, GA4Base):
                 all_selectable_metrics.remove(field)
         return all_selectable_dimensions, all_selectable_metrics
 
-
     def select_random_fields(self):
 
         selected_dimensions = set()
@@ -127,7 +89,8 @@ class GA4AllFieldsTest(AllFieldsTest, GA4Base):
         all_selectable_metrics = list(self.field_exclusions['METRIC'].keys())
 
         # date is always selected, include its field_exclusions
-        all_selectable_dimensions, all_selectable_metrics = self.exclude_field(all_selectable_dimensions, all_selectable_metrics, 'date', self.DIMENSION)
+        all_selectable_dimensions, all_selectable_metrics = self.exclude_field(
+            all_selectable_dimensions, all_selectable_metrics, 'date', self.DIMENSION)
 
         # select fewer dimensions to increase chance of replicating data
         max_dimensions = 4
@@ -143,20 +106,21 @@ class GA4AllFieldsTest(AllFieldsTest, GA4Base):
             if all_selectable_dimensions and len(selected_dimensions) < max_dimensions:
                 random_dimension = choice(all_selectable_dimensions)
                 selected_dimensions.add(random_dimension)
-                all_selectable_dimensions, all_selectable_metrics = self.exclude_field(all_selectable_dimensions, all_selectable_metrics, random_dimension, self.DIMENSION)
+                all_selectable_dimensions, all_selectable_metrics = self.exclude_field(
+                    all_selectable_dimensions, all_selectable_metrics,
+                    random_dimension, self.DIMENSION)
 
             if all_selectable_metrics and len(selected_metrics) < max_metrics:
                 random_metric = choice(all_selectable_metrics)
                 selected_metrics.add(random_metric)
-                all_selectable_dimensions, all_selectable_metrics = self.exclude_field(all_selectable_dimensions, all_selectable_metrics, random_metric, self.METRIC)
+                all_selectable_dimensions, all_selectable_metrics = self.exclude_field(
+                    all_selectable_dimensions, all_selectable_metrics, random_metric, self.METRIC)
 
         return selected_dimensions | selected_metrics
 
-
     ##########################################################################
-    ### Overridden Methods
+    # Overridden Methods
     ##########################################################################
-
 
     # Removes assertion that each stream syncs records
     # No guarantee random selection will sync records
@@ -179,9 +143,8 @@ class GA4AllFieldsTest(AllFieldsTest, GA4Base):
 
         return sync_record_count
 
-
     ##########################################################################
-    ### Overridden Tests
+    # Overridden Tests
     ##########################################################################
 
     # Tests are redefined because there is no guarantee a random field
@@ -196,7 +159,6 @@ class GA4AllFieldsTest(AllFieldsTest, GA4Base):
 
         self.assertTrue(synced_stream_names.issubset(expected_streams))
 
-
     def test_all_fields_for_streams_are_replicated(self):
         for stream in self.streams_to_test():
             with self.subTest(stream=stream):
@@ -205,24 +167,47 @@ class GA4AllFieldsTest(AllFieldsTest, GA4Base):
                 # gather expectations
                 expected_all_keys = self.selected_fields.get(stream_name, set())
 
-                if self.synced_records.get(stream_name):
-                    # gather results
-                    actual_all_keys_per_record = [set(message['data'].keys()) for message in
-                                                  self.get_upsert_messages_for_stream(self.synced_records, stream_name)]
+                records = [
+                    record['data'] for record in
+                    self.synced_records.get(stream, {}).get('messages', [])
+                    if record.get('action') == 'upsert']
 
-                    for actual_all_keys in actual_all_keys_per_record:
-                        self.assertSetEqual(expected_all_keys, actual_all_keys)
+                if records:
+                    fields_replicated = self.actual_fields.get(stream, set())
 
+                    # verify that all fields are sent to the target
+                    # test the combination of all records
+                    self.assertSetEqual(
+                        fields_replicated, expected_all_keys,
+                        logging=f"verify all fields are replicated for stream {stream}")
 
     ##########################################################################
-    ### Tests To Skip
+    # Tests To Skip
     ##########################################################################
-
 
     @unittest.skip("Random selection doesn't always sync records")
     def test_all_streams_sync_records(self):
         pass
 
-
-    def __init__(self, test_run):
-        super().__init__(test_run)
+#  TODO - Run failed, might have an issue with the way we expect data for some fields
+#   https://jira.talendforge.org/browse/TDL-21467 is for the nth_hour field being other
+#   We should remove nth_hour from random selection until the bug is addressed.
+#  Traceback (most recent call last):
+#    File "/usr/local/share/virtualenvs/tap-ga4/bin/tap-ga4", line 11, in <module>
+#      load_entry_point('tap-ga4', 'console_scripts', 'tap-ga4')()
+#    File "/opt/code/tap-ga4/tap_ga4/__init__.py", line 47, in main
+#       raise e
+#    File "/opt/code/tap-ga4/tap_ga4/__init__.py", line 43, in main
+#       main_impl()
+#    File "/opt/code/tap-ga4/tap_ga4/__init__.py", line 35, in main_impl
+#       sync(client, config, catalog, state)
+#    File "/opt/code/tap-ga4/tap_ga4/sync.py", line 250, in sync
+#       sync_report(client, schema, report, start_date, end_date, request_window_size, state)
+#    File "/opt/code/tap-ga4/tap_ga4/sync.py", line 196, in sync_report
+#      transformer.transform(
+#    File "/home/ubuntu/.virtualenvs/tap-ga4/lib/python3.8/site-packages/singer/transform.py", line 153, in transform
+#      raise SchemaMismatch(self.errors)
+#  singer.transform.SchemaMismatch: Errors during transform
+#   date_hour: data does not match {'type': ['string', 'null'], 'format': 'date-time'}
+#   date_hour_minute: data does not match {'type': ['string', 'null'], 'format': 'date-time'}
+#   nth_hour: data does not match {'selected': True, 'inclusion': 'available', 'type': ['integer', 'null']}